Addition of security measures in views

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 
@@ -67,8 +68,14 @@ def product_detail(request, product_id):
 
     return render(request, 'products/product_detail.html', context)
 
+
+@login_required
 def add_product(request):
     # View to Add product to store
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry you are not able to access that')
+        return redirect(reverse('products'))
+    
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -87,8 +94,14 @@ def add_product(request):
 
     return render(request, template, context)
 
+
+@login_required()
 def edit_product(request, product_id):
     # Edit Product in the Store
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry you are not able to access that')
+        return redirect(reverse('products'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -111,8 +124,13 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required()
 def delete_product(request, product_id):
     # delete Product in the Store
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry you are not able to access that')
+        return redirect(reverse('products'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product Deleted')
